+++ spring-core/src/main/java/org/springframework/util/LinkedCaseInsensitiveMap.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2018 the original author or authors.
+ * Copyright 2002-2020 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -220,8 +220,8 @@ public class LinkedCaseInsensitiveMap<V> implements Map<String, V>, Serializable
 	}
 
 	@Override
-	public boolean equals(Object obj) {
-		return this.targetMap.equals(obj);
+	public boolean equals(@Nullable Object other) {
+		return (this == other || this.targetMap.equals(other));
 	}
 
 	@Override
