+++ spring-core/src/main/java/org/springframework/util/LinkedMultiValueMap.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2019 the original author or authors.
+ * Copyright 2002-2020 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,14 +17,10 @@
 package org.springframework.util;
 
 import java.io.Serializable;
-import java.util.Collection;
 import java.util.LinkedHashMap;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
-
-import org.springframework.lang.Nullable;
 
 /**
  * Simple implementation of {@link MultiValueMap} that wraps a {@link LinkedHashMap},
@@ -39,18 +35,16 @@ import org.springframework.lang.Nullable;
  * @param <K> the key type
  * @param <V> the value element type
  */
-public class LinkedMultiValueMap<K, V> implements MultiValueMap<K, V>, Serializable, Cloneable {
+public class LinkedMultiValueMap<K, V> extends MultiValueMapAdapter<K, V> implements Serializable, Cloneable {
 
 	private static final long serialVersionUID = 3801124242820219131L;
-
-	private final Map<K, List<V>> targetMap;
 
 
 	/**
 	 * Create a new LinkedMultiValueMap that wraps a {@link LinkedHashMap}.
 	 */
 	public LinkedMultiValueMap() {
-		this.targetMap = new LinkedHashMap<>();
+		super(new LinkedHashMap<>());
 	}
 
 	/**
@@ -59,7 +53,7 @@ public class LinkedMultiValueMap<K, V> implements MultiValueMap<K, V>, Serializa
 	 * @param initialCapacity the initial capacity
 	 */
 	public LinkedMultiValueMap(int initialCapacity) {
-		this.targetMap = new LinkedHashMap<>(initialCapacity);
+		super(new LinkedHashMap<>(initialCapacity));
 	}
 
 	/**
@@ -71,125 +65,7 @@ public class LinkedMultiValueMap<K, V> implements MultiValueMap<K, V>, Serializa
 	 * @see #deepCopy()
 	 */
 	public LinkedMultiValueMap(Map<K, List<V>> otherMap) {
-		this.targetMap = new LinkedHashMap<>(otherMap);
-	}
-
-
-	// MultiValueMap implementation
-
-	@Override
-	@Nullable
-	public V getFirst(K key) {
-		List<V> values = this.targetMap.get(key);
-		return (values != null && !values.isEmpty() ? values.get(0) : null);
-	}
-
-	@Override
-	public void add(K key, @Nullable V value) {
-		List<V> values = this.targetMap.computeIfAbsent(key, k -> new LinkedList<>());
-		values.add(value);
-	}
-
-	@Override
-	public void addAll(K key, List<? extends V> values) {
-		List<V> currentValues = this.targetMap.computeIfAbsent(key, k -> new LinkedList<>());
-		currentValues.addAll(values);
-	}
-
-	@Override
-	public void addAll(MultiValueMap<K, V> values) {
-		for (Entry<K, List<V>> entry : values.entrySet()) {
-			addAll(entry.getKey(), entry.getValue());
-		}
-	}
-
-	@Override
-	public void set(K key, @Nullable V value) {
-		List<V> values = new LinkedList<>();
-		values.add(value);
-		this.targetMap.put(key, values);
-	}
-
-	@Override
-	public void setAll(Map<K, V> values) {
-		values.forEach(this::set);
-	}
-
-	@Override
-	public Map<K, V> toSingleValueMap() {
-		LinkedHashMap<K, V> singleValueMap = new LinkedHashMap<>(this.targetMap.size());
-		this.targetMap.forEach((key, values) -> {
-			if (values != null && !values.isEmpty()) {
-				singleValueMap.put(key, values.get(0));
-			}
-		});
-		return singleValueMap;
-	}
-
-
-	// Map implementation
-
-	@Override
-	public int size() {
-		return this.targetMap.size();
-	}
-
-	@Override
-	public boolean isEmpty() {
-		return this.targetMap.isEmpty();
-	}
-
-	@Override
-	public boolean containsKey(Object key) {
-		return this.targetMap.containsKey(key);
-	}
-
-	@Override
-	public boolean containsValue(Object value) {
-		return this.targetMap.containsValue(value);
-	}
-
-	@Override
-	@Nullable
-	public List<V> get(Object key) {
-		return this.targetMap.get(key);
-	}
-
-	@Override
-	@Nullable
-	public List<V> put(K key, List<V> value) {
-		return this.targetMap.put(key, value);
-	}
-
-	@Override
-	@Nullable
-	public List<V> remove(Object key) {
-		return this.targetMap.remove(key);
-	}
-
-	@Override
-	public void putAll(Map<? extends K, ? extends List<V>> map) {
-		this.targetMap.putAll(map);
-	}
-
-	@Override
-	public void clear() {
-		this.targetMap.clear();
-	}
-
-	@Override
-	public Set<K> keySet() {
-		return this.targetMap.keySet();
-	}
-
-	@Override
-	public Collection<List<V>> values() {
-		return this.targetMap.values();
-	}
-
-	@Override
-	public Set<Entry<K, List<V>>> entrySet() {
-		return this.targetMap.entrySet();
+		super(new LinkedHashMap<>(otherMap));
 	}
 
 
@@ -203,8 +79,8 @@ public class LinkedMultiValueMap<K, V> implements MultiValueMap<K, V>, Serializa
 	 * @see #clone()
 	 */
 	public LinkedMultiValueMap<K, V> deepCopy() {
-		LinkedMultiValueMap<K, V> copy = new LinkedMultiValueMap<>(this.targetMap.size());
-		this.targetMap.forEach((key, value) -> copy.put(key, new LinkedList<>(value)));
+		LinkedMultiValueMap<K, V> copy = new LinkedMultiValueMap<>(size());
+		forEach((key, values) -> copy.put(key, new LinkedList<>(values)));
 		return copy;
 	}
 
@@ -222,21 +98,6 @@ public class LinkedMultiValueMap<K, V> implements MultiValueMap<K, V>, Serializa
 	@Override
 	public LinkedMultiValueMap<K, V> clone() {
 		return new LinkedMultiValueMap<>(this);
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		return this.targetMap.equals(obj);
-	}
-
-	@Override
-	public int hashCode() {
-		return this.targetMap.hashCode();
-	}
-
-	@Override
-	public String toString() {
-		return this.targetMap.toString();
 	}
 
 }
